Batch: Move `VisGraphWrapper` utility to conftest in tests
