+++ loki/batch/tests/test_scheduler_processing.py
@@ -22,7 +22,7 @@ from loki.frontend import HAVE_FP, available_frontends
 from loki.ir import nodes as ir
 from loki.transformations import DependencyTransformation, ModuleWrapTransformation
 
-from ._test_scheduler_utils import VisGraphWrapper
+from .conftest import VisGraphWrapper
 
 pytestmark = pytest.mark.skipif(not HAVE_FP, reason='Fparser not available')
